feat: add IntoPovider trait and remove connection
