+++ crates/tracel-core/Cargo.toml
@@ -12,7 +12,7 @@ keywords.workspace = true
 categories.workspace = true
 
 [features]
-default = []
+default = ["station"]
 station = ["tracel-client/station"]
 
 [dependencies]
