+++ crates/tracel-core/src/context.rs
@@ -1,11 +1,22 @@
 use std::sync::Arc;
 
-use crate::connection::{Connection, ContextError};
+use crate::backend::cloud::CloudError;
+#[cfg(feature = "station")]
+use crate::backend::station::StationError;
 use crate::dataset::{DatasetModule, DatasetProvider};
 use crate::model_registry::{ModelRegistryModule, ModelRegistryProvider};
 use tracel_experiment::ExperimentModule;
 use tracel_experiment::ExperimentProvider;
 use tracel_inference::{InferenceModule, InferenceProvider};
+
+#[derive(Debug, thiserror::Error)]
+pub enum ContextError {
+    #[error(transparent)]
+    Cloud(#[from] CloudError),
+    #[cfg(feature = "station")]
+    #[error(transparent)]
+    Station(#[from] StationError),
+}
 
 #[derive(Clone)]
 pub struct Context {
@@ -15,8 +26,25 @@ pub struct Context {
     dataset_provider: Option<Arc<dyn DatasetProvider>>,
 }
 
+pub struct Providers {
+    pub experiment: Arc<dyn ExperimentProvider>,
+    pub inference: Arc<dyn InferenceProvider>,
+    pub model_registry: Option<Arc<dyn ModelRegistryProvider>>,
+    pub dataset: Option<Arc<dyn DatasetProvider>>,
+}
+
+pub trait IntoProviders {
+    fn into_providers(self) -> Result<Providers, ContextError>;
+}
+
+impl<T: IntoProviders + 'static> From<T> for Box<dyn IntoProviders> {
+    fn from(backend: T) -> Self {
+        Box::new(backend)
+    }
+}
+
 impl Context {
-    pub fn new(connection: Connection) -> Result<Self, ContextError> {
+    pub fn new(connection: impl IntoProviders) -> Result<Self, ContextError> {
         let providers = connection.into_providers()?;
         Ok(Self {
             experiment_provider: providers.experiment,
