+++ crates/tracel-core/src/lib.rs
@@ -1,5 +1,4 @@
 mod backend;
-mod connection;
 mod context;
 mod dataset;
 mod model_registry;
@@ -7,8 +6,7 @@ mod model_registry;
 pub mod experiment;
 pub mod inference;
 
-pub use connection::{Connection, ContextError};
-pub use context::Context;
+pub use context::{Context, ContextError};
 pub use dataset::{
     DatasetError, DatasetItem, DatasetModule, DatasetVersionSpec, DownloadedDataset,
     StreamedDataset,
